server/requirements.txt
bottle==0.12.18
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
future==0.18.2
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
moviepy==1.0.1
numpy==1.18.1
Pillow==7.0.0
proglog==0.1.9
pycparser==2.19
python-telegram-bot==12.3.0
requests==2.22.0
six==1.13.0
tornado==6.0.3
tqdm==4.41.1
urllib3==1.25.7
